+++ LIBFT/libft.h
@@ -29,7 +29,7 @@ int					ft_tolower(int c);
 
 int					ft_atoi(const char *str);
 
-int					ft_strlen(const char *str);
+size_t					ft_strlen(const char *str);
 size_t				ft_strlcpy(char *dest, const char *src, size_t size);
 size_t				ft_strlcat(char *dest, const char *src, size_t size);
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
